Fix various implicit conversions from size_t to smaller types

See merge request GNOME/glib!4023

+++ gio/gdbusconnection.c
@@ -3822,8 +3822,8 @@ static gboolean
 namespace_rule_matches (const gchar *namespace,
                         const gchar *name)
 {
-  gint len_namespace;
-  gint len_name;
+  size_t len_namespace;
+  size_t len_name;
 
   len_namespace = strlen (namespace);
   len_name = strlen (name);
@@ -3841,7 +3841,7 @@ static gboolean
 path_rule_matches (const gchar *path_a,
                    const gchar *path_b)
 {
-  gint len_a, len_b;
+  size_t len_a, len_b;
 
   len_a = strlen (path_a);
   len_b = strlen (path_b);
